LayoutTests/imported/w3c/web-platform-tests/css/css-viewport/animations/zoom-interpolation-expected.txt


FAIL CSS Transitions: property <zoom> from neutral to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <zoom> from neutral to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from neutral to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <zoom> from neutral to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <zoom> from neutral to [2] at (1) should be [2]
FAIL CSS Transitions: property <zoom> from neutral to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from neutral to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from neutral to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from neutral to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from neutral to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from neutral to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from neutral to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Animations: property <zoom> from neutral to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS CSS Animations: property <zoom> from neutral to [2] at (0) should be [1]
FAIL CSS Animations: property <zoom> from neutral to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL CSS Animations: property <zoom> from neutral to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL CSS Animations: property <zoom> from neutral to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from neutral to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL Web Animations: property <zoom> from neutral to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS Web Animations: property <zoom> from neutral to [2] at (0) should be [1]
FAIL Web Animations: property <zoom> from neutral to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL Web Animations: property <zoom> from neutral to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL Web Animations: property <zoom> from neutral to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from neutral to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL CSS Transitions: property <zoom> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <zoom> from [initial] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <zoom> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <zoom> from [initial] to [2] at (1) should be [2]
FAIL CSS Transitions: property <zoom> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Animations: property <zoom> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS CSS Animations: property <zoom> from [initial] to [2] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL CSS Animations: property <zoom> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL CSS Animations: property <zoom> from [initial] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL Web Animations: property <zoom> from [initial] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS Web Animations: property <zoom> from [initial] to [2] at (0) should be [1]
FAIL Web Animations: property <zoom> from [initial] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL Web Animations: property <zoom> from [initial] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL Web Animations: property <zoom> from [initial] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [initial] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL CSS Transitions: property <zoom> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "2 "
FAIL CSS Transitions: property <zoom> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "2 "
FAIL CSS Transitions: property <zoom> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "2 "
FAIL CSS Transitions: property <zoom> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "2 "
PASS CSS Transitions: property <zoom> from [inherit] to [2] at (1) should be [2]
FAIL CSS Transitions: property <zoom> from [inherit] to [2] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [inherit] to [2] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "2 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "1 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "1 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "1 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "1 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [inherit] to [2] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (-0.3) should be [3.3] assert_equals: expected "3.3 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (0) should be [3] assert_equals: expected "3 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (0.3) should be [2.7] assert_equals: expected "2.7 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (0.6) should be [2.4] assert_equals: expected "2.4 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [inherit] to [2] at (1.5) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL CSS Transitions: property <zoom> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <zoom> from [unset] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <zoom> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <zoom> from [unset] to [2] at (1) should be [2]
FAIL CSS Transitions: property <zoom> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Animations: property <zoom> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS CSS Animations: property <zoom> from [unset] to [2] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL CSS Animations: property <zoom> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL CSS Animations: property <zoom> from [unset] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL Web Animations: property <zoom> from [unset] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS Web Animations: property <zoom> from [unset] to [2] at (0) should be [1]
FAIL Web Animations: property <zoom> from [unset] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL Web Animations: property <zoom> from [unset] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL Web Animations: property <zoom> from [unset] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [unset] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL CSS Transitions: property <zoom> from [1] to [2] at (-5) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions: property <zoom> from [1] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions: property <zoom> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions: property <zoom> from [1] to [2] at (1) should be [2]
FAIL CSS Transitions: property <zoom> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (-5) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [1] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "2 "
PASS CSS Animations: property <zoom> from [1] to [2] at (-5) should be [1]
FAIL CSS Animations: property <zoom> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS CSS Animations: property <zoom> from [1] to [2] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL CSS Animations: property <zoom> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL CSS Animations: property <zoom> from [1] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
PASS Web Animations: property <zoom> from [1] to [2] at (-5) should be [1]
FAIL Web Animations: property <zoom> from [1] to [2] at (-0.3) should be [0.7] assert_equals: expected "0.7 " but got "1 "
PASS Web Animations: property <zoom> from [1] to [2] at (0) should be [1]
FAIL Web Animations: property <zoom> from [1] to [2] at (0.3) should be [1.3] assert_equals: expected "1.3 " but got "1 "
FAIL Web Animations: property <zoom> from [1] to [2] at (0.6) should be [1.6] assert_equals: expected "1.6 " but got "1 "
FAIL Web Animations: property <zoom> from [1] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [1] to [2] at (1.5) should be [2.5] assert_equals: expected "2.5 " but got "1 "
FAIL CSS Transitions: property <zoom> from [0.5] to [1.5] at (-5) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL CSS Transitions: property <zoom> from [0.5] to [1.5] at (-0.3) should be [0.2] assert_equals: expected "0.2 " but got "1.5 "
FAIL CSS Transitions: property <zoom> from [0.5] to [1.5] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1.5 "
FAIL CSS Transitions: property <zoom> from [0.5] to [1.5] at (0.5) should be [1] assert_equals: expected "1 " but got "1.5 "
PASS CSS Transitions: property <zoom> from [0.5] to [1.5] at (1) should be [1.5]
FAIL CSS Transitions: property <zoom> from [0.5] to [1.5] at (1.5) should be [2] assert_equals: expected "2 " but got "1.5 "
FAIL CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (-5) should be [1] assert_equals: expected "1 " but got "1.5 "
FAIL CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (-0.3) should be [0.2] assert_equals: expected "0.2 " but got "1.5 "
FAIL CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1.5 "
FAIL CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (0.5) should be [1] assert_equals: expected "1 " but got "1.5 "
PASS CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (1) should be [1.5]
FAIL CSS Transitions with transition: all: property <zoom> from [0.5] to [1.5] at (1.5) should be [2] assert_equals: expected "2 " but got "1.5 "
PASS CSS Animations: property <zoom> from [0.5] to [1.5] at (-5) should be [1]
FAIL CSS Animations: property <zoom> from [0.5] to [1.5] at (-0.3) should be [0.2] assert_equals: expected "0.2 " but got "1 "
FAIL CSS Animations: property <zoom> from [0.5] to [1.5] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1 "
PASS CSS Animations: property <zoom> from [0.5] to [1.5] at (0.5) should be [1]
FAIL CSS Animations: property <zoom> from [0.5] to [1.5] at (1) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL CSS Animations: property <zoom> from [0.5] to [1.5] at (1.5) should be [2] assert_equals: expected "2 " but got "1 "
PASS Web Animations: property <zoom> from [0.5] to [1.5] at (-5) should be [1]
FAIL Web Animations: property <zoom> from [0.5] to [1.5] at (-0.3) should be [0.2] assert_equals: expected "0.2 " but got "1 "
FAIL Web Animations: property <zoom> from [0.5] to [1.5] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1 "
PASS Web Animations: property <zoom> from [0.5] to [1.5] at (0.5) should be [1]
FAIL Web Animations: property <zoom> from [0.5] to [1.5] at (1) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL Web Animations: property <zoom> from [0.5] to [1.5] at (1.5) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Transitions: property <zoom> from [normal] to [3] at (-0.5) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions: property <zoom> from [normal] to [3] at (0) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions: property <zoom> from [normal] to [3] at (0.5) should be [2] assert_equals: expected "2 " but got "3 "
PASS CSS Transitions: property <zoom> from [normal] to [3] at (1) should be [3]
FAIL CSS Transitions with transition: all: property <zoom> from [normal] to [3] at (-0.5) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions with transition: all: property <zoom> from [normal] to [3] at (0) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions with transition: all: property <zoom> from [normal] to [3] at (0.5) should be [2] assert_equals: expected "2 " but got "3 "
PASS CSS Transitions with transition: all: property <zoom> from [normal] to [3] at (1) should be [3]
PASS CSS Animations: property <zoom> from [normal] to [3] at (-0.5) should be [1]
PASS CSS Animations: property <zoom> from [normal] to [3] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [normal] to [3] at (0.5) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [normal] to [3] at (1) should be [3] assert_equals: expected "3 " but got "1 "
PASS Web Animations: property <zoom> from [normal] to [3] at (-0.5) should be [1]
PASS Web Animations: property <zoom> from [normal] to [3] at (0) should be [1]
FAIL Web Animations: property <zoom> from [normal] to [3] at (0.5) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [normal] to [3] at (1) should be [3] assert_equals: expected "3 " but got "1 "
FAIL CSS Transitions: property <zoom> from [100%] to [300%] at (0) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions: property <zoom> from [100%] to [300%] at (0.5) should be [2] assert_equals: expected "2 " but got "3 "
PASS CSS Transitions: property <zoom> from [100%] to [300%] at (1) should be [3]
FAIL CSS Transitions with transition: all: property <zoom> from [100%] to [300%] at (0) should be [1] assert_equals: expected "1 " but got "3 "
FAIL CSS Transitions with transition: all: property <zoom> from [100%] to [300%] at (0.5) should be [2] assert_equals: expected "2 " but got "3 "
PASS CSS Transitions with transition: all: property <zoom> from [100%] to [300%] at (1) should be [3]
PASS CSS Animations: property <zoom> from [100%] to [300%] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [100%] to [300%] at (0.5) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Animations: property <zoom> from [100%] to [300%] at (1) should be [3] assert_equals: expected "3 " but got "1 "
PASS Web Animations: property <zoom> from [100%] to [300%] at (0) should be [1]
FAIL Web Animations: property <zoom> from [100%] to [300%] at (0.5) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [100%] to [300%] at (1) should be [3] assert_equals: expected "3 " but got "1 "
FAIL CSS Transitions: property <zoom> from [50%] to [2] at (0) should be [0.5] assert_equals: expected "0.5 " but got "2 "
FAIL CSS Transitions: property <zoom> from [50%] to [2] at (0.5) should be [1.25] assert_equals: expected "1.25 " but got "2 "
PASS CSS Transitions: property <zoom> from [50%] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [50%] to [2] at (0) should be [0.5] assert_equals: expected "0.5 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [50%] to [2] at (0.5) should be [1.25] assert_equals: expected "1.25 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [50%] to [2] at (1) should be [2]
FAIL CSS Animations: property <zoom> from [50%] to [2] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1 "
FAIL CSS Animations: property <zoom> from [50%] to [2] at (0.5) should be [1.25] assert_equals: expected "1.25 " but got "1 "
FAIL CSS Animations: property <zoom> from [50%] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Web Animations: property <zoom> from [50%] to [2] at (0) should be [0.5] assert_equals: expected "0.5 " but got "1 "
FAIL Web Animations: property <zoom> from [50%] to [2] at (0.5) should be [1.25] assert_equals: expected "1.25 " but got "1 "
FAIL Web Animations: property <zoom> from [50%] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL CSS Transitions: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0.5) should be [1.5] assert_equals: expected "1.5 " but got "2 "
PASS CSS Transitions: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (1) should be [2]
FAIL CSS Transitions with transition: all: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0) should be [1] assert_equals: expected "1 " but got "2 "
FAIL CSS Transitions with transition: all: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0.5) should be [1.5] assert_equals: expected "1.5 " but got "2 "
PASS CSS Transitions with transition: all: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (1) should be [2]
PASS CSS Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0) should be [1]
FAIL CSS Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0.5) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL CSS Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
PASS Web Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0) should be [1]
FAIL Web Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (0.5) should be [1.5] assert_equals: expected "1.5 " but got "1 "
FAIL Web Animations: property <zoom> from [calc(sibling-index() * 100%)] to [2] at (1) should be [2] assert_equals: expected "2 " but got "1 "
FAIL Animating zoom affects layout using the interpolated value assert_approx_equals: expected 1.5 +/- 0.01 but got 1
FAIL transition: all animates zoom changes from :hover assert_approx_equals: expected 1.5 +/- 0.01 but got 2
